basic documentation comments parser
A parser of Java documentation comments. Documentation comments are
represented as a list of spans. A span is either a plain text or a tag.
Tags can be inline or block. A comment is divided into a main description
and a tag section. The main description part contains text spans and inline
tags. The tag section contains whitespace text spans and block tags. Tags have
a name and content. Content of an inline tag is a text, content of a block tag
is a list of text spans and inline tags. In addition to that, the first
sentence of main description can be retrieved separately.

For now, various tag types can only be distinguished by their names;
there will be special interfaces for known tag types in the future.